refactor!: make solutions a submodule and break map class out to its own file
